+++ publications/__init__.py
@@ -1,4 +1,0 @@
-__license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
-__authors__ = ['Lucas Theis <lucas@theis.io>', 'Marc Bourqui']
-__docformat__ = 'epytext'
-__version__ = '0.7.0-dev'