+++ publications_bootstrap/__init__.py
@@ -1,0 +1,7 @@
+__license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
+__authors__ = ['Lucas Theis <lucas@theis.io>', 'Marc Bourqui <https://github.com/mbourqui>']
+__docformat__ = 'epytext'
+__version__ = '1.0.0'
+__version_info__ = tuple(
+    [int(num) if num.isdigit() else num for num in __version__.replace('-', '.', 1).split('.')])
+__status__ = 'Stable'